Convert 'minimal duration' param to real (double), which avoids changing +- 1 when exposing to QML. Adjusted in Tglobals and TcommonListener. Get rid of 'play detected' parameter - dead code

+++ src/libs/sound/tcommonlistener.cpp
@@ -87,13 +87,13 @@ void TcommonListener::setAudioInParams() {
 }
 
 
-void TcommonListener::setMinimalVolume(float minVol) {
-  finder()->setMinimalVolume(minVol);
+void TcommonListener::setMinimalVolume(qreal minVol) {
+  finder()->setMinimalVolume(static_cast<float>(minVol));
   m_audioParams->minimalVol = minVol;
 }
 
 
-float TcommonListener::minimalVolume() {
+qreal TcommonListener::minimalVolume() {
   return m_audioParams->minimalVol;
 }
